moved major part of "/" to server
